requirements.txt
torch==2.3.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
torchvision==0.18.0
fastapi==0.111.0
uvicorn==0.30.1